Use base integration tester for language plugins default source set
